requirements.txt
Flask>=3.0.0
Flask-WTF>=1.2.1
python-dotenv>=1.0.1
pytest>=8.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0
